api create equipment

+++ app/Jobs/UploadImageToStorage.php
@@ -4,6 +4,7 @@ namespace App\Jobs;
 
 use App\Helpers\FileUtils;
 use App\Helpers\S3Utils;
+use App\Models\Equipment;
 use App\Models\Feedback;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Queue\Queueable;
@@ -31,26 +32,56 @@ class UploadImageToStorage implements ShouldQueue
     /**
      * Execute the job.
      */
-    public function handle(): void
-    {
-        switch ($this->_type) {
-            case config('constant.type.feedback'): {
-                $feedback = Feedback::find($this->_objectId);
-                $images = [];
-                foreach ($this->_images as $image) {
-                    $fileName = now()->format('Y-m-d') . "/" . $this->_objectId . "/" . uniqid();
-                    Log::info($image);
-                    $file = FileUtils::convertBase64ToFile($image, 'image');
-                    $url = S3Utils::upload($file, $fileName, 'feedback');
-                    unlink($file->getPathname());
-                    $images[] = $url;
-                }
+    public function handle(): void{
 
-                $body = $feedback->body;
-                $body['images'] = $images;
-                $feedback->body = $body;
-                $feedback->save();
-            } break;
-        }
+    switch ($this->_type) {
+        case config('constant.type.feedback'):
+            $feedback = Feedback::find($this->_objectId);
+            if (!$feedback) {
+                return;
+            }
+
+            $images = $this->uploadImages($this->_images, 'feedback');
+            $body = $feedback->body;
+            $body['images'] = $images;
+            $feedback->body = $body;
+            $feedback->save();
+            break;
+
+        case config('constant.type.equipment'):
+            $equipment = Equipment::find($this->_objectId);
+            if (!$equipment || empty($this->_images)) {
+                return;
+            }
+
+            $url = $this->uploadImage($this->_images, 'equipment');
+            $equipment->thumbnail = $url;
+            $equipment->save();
+            break;
+    }
+}
+
+    /**
+     * Hàm xử lý nhiều ảnh (cho feedback)
+     */
+    private function uploadImages(array $images, string $folder): array
+    {
+        return array_map(fn($image) => $this->uploadImage($image, $folder), array_filter($images));
+    }
+
+    /**
+     * Hàm xử lý 1 ảnh duy nhất
+     */
+    private function uploadImage(string $image, string $folder): ?string
+    {
+        if (!$image) return null;
+
+        $fileName = now()->format('Y-m-d') . "/" . $this->_objectId . "/" . uniqid();
+        $file = FileUtils::convertBase64ToFile($image, 'image');
+
+        $url = S3Utils::upload($file, $fileName, $folder);
+        unlink($file->getPathname());
+
+        return $url;
     }
 }
